1. Errors, CodeStyle and architecture fixes before merge
